+++ tests/test_data.py
@@ -62,7 +62,7 @@ def test_get_batch():
                 f"Starting index {starting_index} occurs {count} times, but expected at most {occurrences_upper_bound}"
             )
 
-    with pytest.raises(RuntimeError) as excinfo:
+    with pytest.raises((RuntimeError, AssertionError)) as excinfo:
         # We're assuming that cuda:99 is an invalid device ordinal.
         # Just adding this here to make sure that the device flag is
         # being handled.
@@ -72,4 +72,6 @@ def test_get_batch():
             context_length=context_length,
             device="cuda:99",
         )
-        assert "CUDA error" in str(excinfo.value)
+        assert "CUDA error" in str(
+            excinfo.value
+        ) or "Torch not compiled with CUDA enabled" in str(excinfo.value)