+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All changes we make to the assignment code or PDF will be documented in this fil
 
 ### Fixed
 
+- code: fix `test_get_batch` to handle "AssertionError: Torch not compiled with CUDA enabled". 
+
 ## [0.1.6] - 2024-04-13
 
 ### Added